Added database.yml
This may not be such a good idea, however I have had issues in the past when the database.yml was missing. The Rails Tutorial put the database.yml in the gitignore, likely because the production used postgres, however I think that was unnecessary since when deploying to heroku it indicated it was writing to the database.yml file to make it work on heroku.