Remove redirect on media control
Removed redirect when using media control buttons in header, but to use it requires javascript. Was unable to implement unobtrusive javascript, so if javascript is disabled, then media control will not work. Added song list link as an mp3 player icon since it became inaccessible when the redirects were removed.